requirements.txt
tensorflow-gpu==1.11
keras==2.2.4
numpy==1.16.4
flask==1.1.1
scipy
Pillow
cython
matplotlib
scikit-image
h5py
imgaug
IPython[all]
